Fix safe money being wiped
This is a bad solution to prevent duping because it wipes safe money often. When server_updateObject runs before the unlocked safe (_obj) is deleted it updates the safe in the database as having zero coins. That happens for objects within 10m when server_playerSync is called (which is several times per second), so there is a good chance it will occur while this script is running. It happened to me almost every time on my local test server. As soon as I removed this line all the safe money saved again.